Renamed options files for clarity; added separate file for baseline feed data modifications; added author_byline field to post data
